requirements.txt
gliner==0.1.7
transformers==4.39.3
autocorrect==2.6.1
streamlit==1.33.0
nltk==3.8.1
langchain==0.1.16
scipy==1.10.1
langchain-openai
langchainhub
sentence-transformers
chromadb
lark
mysql-connector-python
python-dotenv